replace 'cf-ipcountry' with Swetrix own implementation
