+++ apps/production/src/auth/auth.module.ts
@@ -1,11 +1,11 @@
 import { forwardRef, Module } from '@nestjs/common'
 import { JwtModule } from '@nestjs/jwt'
 import { PassportModule } from '@nestjs/passport'
-import { ActionTokensModule } from 'src/action-tokens/action-tokens.module'
-import { MailerModule } from 'src/mailer/mailer.module'
-import { UserModule } from 'src/user/user.module'
-import { ProjectModule } from 'src/project/project.module'
-import { TelegramService } from 'src/integrations/telegram/telegram.service'
+import { ActionTokensModule } from '../action-tokens/action-tokens.module'
+import { MailerModule } from '../mailer/mailer.module'
+import { UserModule } from '../user/user.module'
+import { ProjectModule } from '../project/project.module'
+import { TelegramService } from '../integrations/telegram/telegram.service'
 import { AuthController } from './auth.controller'
 import { AuthService } from './auth.service'
 import {
